+++ packages/elixir/lumis/lib/lumis.ex
@@ -895,9 +895,9 @@ defmodule Lumis do
     case Lumis.Native.highlight(source, options) do
       {:error, {:language_not_loaded, language}} ->
         {:error,
-         "language #{inspect(language)} could not be loaded. Cache it ahead of " <>
-           "time with `mix lumis.languages.cache #{language}` if this host has " <>
-           "no network access"}
+         "language #{inspect(language)} could not be loaded. Prepare it with " <>
+           "`Lumis.Languages.cache([#{inspect(language)}])` during application " <>
+           "startup if this host has no network access"}
 
       other ->
         other